--- packages/soroban/contracts/reviews-ratings/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "reviews-ratings"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-doctest = false
-
-[dependencies]
-soroban-sdk = { workspace = true }
-
-[dev-dependencies]
-soroban-sdk = { workspace = true, features = ["testutils"] }
